+++ api/README.md
@@ -1,0 +1,3 @@
+Kai API
+===========
+This module contains The REST api deployed as part of Kai
