+++ platform/README.md
@@ -1,0 +1,5 @@
+Kai Platform
+==================
+
+The Kai Platform contains the stacks and  constructs which create the 
+underlying platform which hosts user's graphs
